% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vector.R
\name{add_vector}
\alias{add_vector}
\title{Add a vector
Computes the log map vector for a new tree from a base tree and adds it to an existing
set of log map vectors.}
\usage{
add_vector(
  new_tree_path,
  base_path,
  vectors,
  add_tip_branches = TRUE,
  new_name = NULL
)
}
\arguments{
\item{new_tree_path}{The path to the new tree to add.}

\item{base_path}{Path to the base tree.}

\item{vectors}{Set of log map vectors for original set of trees.}

\item{add_tip_branches}{A boolean. When true, the lengths of branches ending in tips will be added
to the log map coordinates. Default to TRUE.}

\item{new_name}{Optional name of new tree to be added.}
}
\value{
A set of log map vectors with a new vector added for the new tree.
}
\description{
Add a vector
Computes the log map vector for a new tree from a base tree and adds it to an existing
set of log map vectors.
}
\examples{
path <- paste0(system.file("txt", package = "groves"), "/")
tree_names <- c("tree1", "tree2", "tree3")
tree_paths <- paste0(path, "tree", 1:3, ".txt")
lm_res <- compute_logmap(tree_paths = tree_paths, tree_names = tree_names)
new <- paste0(path, "tree4.txt")
base <- tree_paths[which(tree_names \%in\% lm_res$base_lab)]
add_vector(new_tree_path = new, base_path = base, vectors = lm_res$vectors,
           new_name = "tree4")

}
