% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_align_to_gene_align.R
\name{cat_align_to_gene_align}
\alias{cat_align_to_gene_align}
\title{Creates gene alignments
Creates individual gene alignments from a concatenated alignment.}
\usage{
cat_align_to_gene_align(
  concat_path,
  gene_names = NULL,
  sep_string = "XXXXX",
  output_path = "gene_alignments/",
  tail = ".faa"
)
}
\arguments{
\item{concat_path}{Path to concatenated alignment.}

\item{gene_names}{A vector of gene names, ordered in the same way in which 
genes are concatenated in the concatenated alignment. If not provided, genes 
will be given numbers.}

\item{sep_string}{The string that separates each gene alignment within the 
concatenated alignment. Defaults to "XXXXX".}

\item{output_path}{The path to a folder in which the gene alignments will be saved. 
Make sure to add a "/" at the end of the path.}

\item{tail}{The end of the gene alignment name. Defaults to ".faa".}
}
\value{
Nothing, creates a folder in the desired directory that includes the 
individual gene alignments.
}
\description{
Creates gene alignments
Creates individual gene alignments from a concatenated alignment.
}
