% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_binary_single.R
\name{check_binary_single}
\alias{check_binary_single}
\title{Check if single tree is binary 
Given a single tree, checks if it is binary, i.e. for a rooted tree with \code{n} tips it has \code{n - 1} internal nodes, 
for a unrooted tree with \code{n} tips it has \code{n - 2} internal nodes, and all internal edges have length 
greater than \code{0}.}
\usage{
check_binary_single(tree)
}
\arguments{
\item{tree}{the tree to check.}
}
\value{
A boolean, TRUE if the tree is binary.
}
\description{
Check if single tree is binary 
Given a single tree, checks if it is binary, i.e. for a rooted tree with \code{n} tips it has \code{n - 1} internal nodes, 
for a unrooted tree with \code{n} tips it has \code{n - 2} internal nodes, and all internal edges have length 
greater than \code{0}.
}
