% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gene_support.R
\name{check_gene_support}
\alias{check_gene_support}
\title{Check gene support
Check the proportion of trees that support splits in a single tree.}
\usage{
check_gene_support(main_tree, trees, rooted = FALSE)
}
\arguments{
\item{main_tree}{A tree to check support for.}

\item{trees}{A set of trees to use to check support for main_tree.}

\item{rooted}{TRUE if trees are rooted, FALSE otherwise. Defaults to FALSE.}
}
\value{
A vector of support proportion for each internal node of \code{main_tree}.
}
\description{
Check gene support
Check the proportion of trees that support splits in a single tree.
}
\examples{
path1 <- system.file("txt", "gene_trees.txt", package = "groves")
path2 <- system.file("txt", "concat_tree.txt", package = "groves")
gene_trees <- ape::read.tree(path1)
concat_tree <- ape::read.tree(path2)
check_gene_support(concat_tree, gene_trees, rooted = FALSE)

}
