% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missingness.R
\name{check_missingness}
\alias{check_missingness}
\title{Check missingness
Compute the proportion of missingness for each genome and gene combination.}
\usage{
check_missingness(
  gene_names,
  tip_names,
  path = "",
  tail = ".fa",
  missing_char = "-"
)
}
\arguments{
\item{gene_names}{A list of names of genes.}

\item{tip_names}{A list of names of genomes (tips of the phylogeny).}

\item{path}{The path to the fasta file.}

\item{tail}{The end of the path to the fasta file.}

\item{missing_char}{The character used in the alignment to denote a missing base. "-" by default.}
}
\value{
A matrix with the proportion of missingness for each genome and gene combination.
}
\description{
Check missingness
Compute the proportion of missingness for each genome and gene combination.
}
\examples{
genes <- c("BacA", "CorA")
tips <- c("GCA_000025925.1", "GCA_000144405.1")
path <- paste0(system.file("faa/", package = "groves"), "/")
miss <- check_missingness(gene_names = genes,
                         tip_names = tips, 
                         path = path, 
                         tail = "_aln.faa")

}
