% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_support.R
\name{check_support}
\alias{check_support}
\title{Check support
Check the proportion of trees that support splits in a single tree.}
\usage{
check_support(main_tree, trees)
}
\arguments{
\item{main_tree}{A tree to check support for.}

\item{trees}{A set of trees to use to check support for main_tree.}
}
\value{
A list with an average support value (averaged over all splits in all trees) and
a support proportion for each internal branch of main_tree.
}
\description{
Check support
Check the proportion of trees that support splits in a single tree.
}
\examples{
path1 <- system.file("txt", "gene_trees.txt", package = "groves")
path2 <- system.file("txt", "concat_tree.txt", package = "groves")
gene_trees <- ape::read.tree(path1)
concat_tree <- ape::read.tree(path2)
check_support(concat_tree, gene_trees)

}
