% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_geodesic.R
\name{compute_geodesic}
\alias{compute_geodesic}
\title{Geodesic implementation
Calculates the geodesic distance between each tree}
\usage{
compute_geodesic(tree_path)
}
\arguments{
\item{tree_path}{The file path to a txt file containing phylogenetic trees in a multiPhylo object.}
}
\value{
A matrix containing distances between each tree in the multiPhylo object. Self distances are set to 0.
}
\description{
Geodesic implementation
Calculates the geodesic distance between each tree
}
\examples{
compute_geodesic(system.file("txt", "small_tree_set.txt", package = "groves"))

}
