% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_logmap.R
\name{compute_logmap}
\alias{compute_logmap}
\title{Compute logmap
Compute the logmap centered at one tree for all other trees.}
\usage{
compute_logmap(
  tree_paths,
  add_tip_branches = TRUE,
  tree_names = NULL,
  base_lab = NULL
)
}
\arguments{
\item{tree_paths}{A list of file paths to all other trees.}

\item{add_tip_branches}{A boolean. When true, the lengths of branches ending in tips will be added
to the log map coordinates. Default to TRUE.}

\item{tree_names}{An optional vector of names of trees. This should be the same length and order
as tree_paths. If base_in_tree_paths = FALSE, this vector should not include the base tree name. 
If this is not given, trees in tree_paths will be named numerically based on their order.}

\item{base_lab}{An optional label to specify the base tree. If this argument is ignored, \code{groves}
will set the base tree to be the tree in the tree set with the smallest mean BHV distance from the
other trees. If included, base_lab can either be a name included in \code{tree_names} or a number 
corresponding with the desired tree's order in \code{tree_paths}.}
}
\value{
A matrix with log map coordinates for all trees. Row names give the name of each tree, or
the order from tree_paths if tree_names are not provided.
}
\description{
Compute logmap
Compute the logmap centered at one tree for all other trees.
}
\examples{
path <- paste0(system.file("txt", package = "groves"), "/")
compute_logmap(tree_paths = paste0(path, "tree", 1:3, ".txt"),
               tree_names = c("tree1", "tree2", "tree3"))

}
