% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_MDS.R
\name{compute_MDS}
\alias{compute_MDS}
\title{MDS on trees
Calculates the distance between each tree and performs MDS}
\usage{
compute_MDS(
  trees_path = NULL,
  nf = 2,
  dist_metric = "BHV",
  dist_matrix = NULL,
  mds_type = "metric",
  tree_names = NULL
)
}
\arguments{
\item{trees_path}{The file path to a txt file containing the set of phylogenetic trees in a 
multiPhylo object, required unless a dist_matrix is not NULL.}

\item{nf}{The number of axes in the MDS to keep. Defaults to 2.}

\item{dist_metric}{Either 'RF' for Robinson-Foulds, or 'BHV' for the geodesic defined in 
Billera et al. (2001). Defaults to 'BHV'. Ignore this argument if using argument dist_matrix.}

\item{dist_matrix}{If you would like to submit your own distance matrix computed with a different
metric, use this argument. In this case, do not include the dist_metric argument.}

\item{mds_type}{MDS is performed with the package smacof. Default is metric scaling, otherwise
use value "nonmetric".}

\item{tree_names}{Optional tree names to use.}
}
\value{
A list containing an object 'dist_mat' containing distances between each phylogenetic tree
in the multiPhylo object, and an object 'df' containing the desired number of axes in the MDS of the trees.
}
\description{
MDS on trees
Calculates the distance between each tree and performs MDS
}
\examples{
trees_path <- system.file("txt", "small_tree_set.txt", package = "groves")
compute_MDS(trees_path = trees_path, tree_names = paste0("tree", 1:3))


}
