% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_topologies.R
\name{count_topologies}
\alias{count_topologies}
\title{Count topologies
Count number of distinct topologies present in group of phylogenetic trees on the same set of
tips.}
\usage{
count_topologies(trees, names = NULL)
}
\arguments{
\item{trees}{A set of trees.}

\item{names}{An optional set of tree names.}
}
\value{
A list containing the number of distinct topologies and a dataframe with topology
information.
}
\description{
Count topologies
Count number of distinct topologies present in group of phylogenetic trees on the same set of
tips.
}
\examples{
gene_trees <- read.tree("data/gene_trees.txt")
count_topologies(gene_trees)

}
