% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_missing_outgroup.R
\name{genes_missing_outgroup}
\alias{genes_missing_outgroup}
\title{Finds missing genes
Provides a subset of genes from a list that are missing in the outgroup.}
\usage{
genes_missing_outgroup(presence, outgroup)
}
\arguments{
\item{presence}{Matrix of presence data for genes and genomes (an entry is TRUE if row gene is
present in column genome).}

\item{outgroup}{Name of the outgroup genome.}
}
\value{
A subset of gene_names that are missing in the outgroup genome
}
\description{
Finds missing genes
Provides a subset of genes from a list that are missing in the outgroup.
}
\examples{
gene_set <- paste0("gene_", 1:3)
faa_path <- paste0(system.file("faa/", package = "groves"), "/")
pres <- get_presence(gene_names = gene_set, path = faa_path, tail = ".faa")
genes_missing_outgroup(pres, "tip_3")

}
