% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence.R
\name{get_presence}
\alias{get_presence}
\title{Makes presence/absence matrix 
Takes a list of genes and returns a presence/absence matrix for genomes that included
as tips in the alignment for each gene.}
\usage{
get_presence(gene_names, path = "", tail = ".fa", extra_col = 10)
}
\arguments{
\item{gene_names}{A list of gene names that correspond with alignments to examine}

\item{path}{The file path to the gene alignments, default is ""}

\item{tail}{The extension to the gene alignments, default is ".fa"}

\item{extra_col}{The number of extra columns added to presence matrix to account for 
tips that are in some genes but not others. Defaults to 10.}
}
\value{
A matrix with rows for each gene and columns for each genome (tip). Entries 
equal to 1 mean the gene alignment includes that tip.
}
\description{
Makes presence/absence matrix 
Takes a list of genes and returns a presence/absence matrix for genomes that included
as tips in the alignment for each gene.
}
\examples{
gene_set <- paste0("gene_", 1:3)
faa_path <- paste0(system.file("faa/", package = "groves"), "/")
get_presence(gene_names = gene_set, path = faa_path, tail = ".faa")

}
