% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_long_branch_tips.R
\name{list_long_branch_tips}
\alias{list_long_branch_tips}
\title{List long branch tips
Returns a list of tips that are children of the longest branch}
\usage{
list_long_branch_tips(tree)
}
\arguments{
\item{tree}{A phylo or multiphylo object.}
}
\value{
A list of tips that are children of the longest branch of a single tree (for a phylo object) and
a list of lists of tips that are children of the longest branch of each tree (for a multiPhylo object).
}
\description{
List long branch tips
Returns a list of tips that are children of the longest branch
}
