% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_txt.R
\name{merge_txt}
\alias{merge_txt}
\title{Merge txt files
Takes the contents of separate txt files and inputs each file as a line in a single file.}
\usage{
merge_txt(paths, output_path)
}
\arguments{
\item{paths}{A set of paths to txt files, each containing a single line.}

\item{output_path}{A path to save the output txt file as.}
}
\description{
Merge txt files
Takes the contents of separate txt files and inputs each file as a line in a single file.
}
\examples{
path <- paste0(system.file("txt", package = "groves"), "/")
tree_paths <- paste0(path, "tree", 1:3, ".txt")
merge_txt(paths = tree_paths, output_path = paste0(path, "test_trees.txt"))

}
