% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_MDS.R
\name{perform_MDS}
\alias{perform_MDS}
\title{Classical MDS of six difference tree distances
Performs Classical MDS for a set of trees using six different distances. Add names and references later.}
\usage{
perform_MDS(trees, tree_names = NULL, consensus = NULL)
}
\arguments{
\item{trees}{A list of phylogenetic trees. If list includes a consensus tree, it should be the
first tree.}

\item{tree_names}{An optional list of gene names to add to the dataframes.}

\item{consensus}{An optional argument specifying whether the first tree is a consensus tree, and
if so the name of the consensus tree.}
}
\value{
A list containing two dataframes. The first, of dimension \eqn{2 \cdot (n_tree\cdot n_dist)},
 contains the first two MDS coordinates for each tree and distance combination. The second, of dimension
 \eqn{(n_tree -1)\cdot n_dist} contains the distances from the first tree to each of the other trees for
 each distance.
}
\description{
Classical MDS of six difference tree distances
Performs Classical MDS for a set of trees using six different distances. Add names and references later.
}
