% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boot_support.R
\name{plot_boot_support}
\alias{plot_boot_support}
\title{Plot bootstrap support
Plot a tree with bootstrap support values on each edge.}
\usage{
plot_boot_support(
  main_tree,
  support,
  lab_size = 2,
  supp_size = 2,
  xlim_max = 0.5,
  hjust = 1
)
}
\arguments{
\item{main_tree}{A tree to plot edge information for.}

\item{support}{A list of values for each branch of main_tree.}

\item{lab_size}{The size of the tip labels, default is 2.}

\item{supp_size}{The size of the support labels, default is 2.}

\item{xlim_max}{The length of the x axis.}

\item{hjust}{Horizontal adjustment for support labels.}
}
\value{
A ggtree object.
}
\description{
Plot bootstrap support
Plot a tree with bootstrap support values on each edge.
}
\examples{
path <- system.file("txt", "tree1.txt", package = "groves")
tree <- ape::read.tree(path)
support <- rep(80, 8)
plot_boot_support(main_tree = tree, support = support, xlim_max = 2.5)

}
