% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_support.R
\name{plot_gene_support}
\alias{plot_gene_support}
\title{Plot gene tree support
Plot a tree with gene tree support values on each edge.}
\usage{
plot_gene_support(
  main_tree,
  support,
  lab_size = 2,
  supp_size = 2,
  xlim_max = 0.5,
  hjust = 1
)
}
\arguments{
\item{main_tree}{A rooted tree to plot edge information for.}

\item{support}{A list of values for each branch of main_tree.}

\item{lab_size}{The size of the tip labels, default is 2.}

\item{supp_size}{The size of the support labels, default is 2.}

\item{xlim_max}{The length of the x axis.}

\item{hjust}{Horizontal adjustment for support labels.}
}
\value{
A ggtree object.
}
\description{
Plot gene tree support
Plot a tree with gene tree support values on each edge.
}
\examples{
path <- system.file("txt", "small_tree_set.txt", package = "groves")
trees <- ape::read.tree(path)
support <- check_support(trees[[1]], trees)$branch_support
plot_gene_support(main_tree = trees[[1]], support = support, xlim_max = 4)

}
