% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MDS.R
\name{plot_MDS}
\alias{plot_MDS}
\title{Plots MDS results
Plots MDS results of trees for each distance.}
\usage{
plot_MDS(
  df,
  consensus = NULL,
  group = NULL,
  title = "MDS of Gene Tree Distances",
  x_dim = "scale_x_dim",
  y_dim = "scale_y_dim",
  show_legend = FALSE,
  legend_lab = NULL,
  scales_arg = "free",
  single_method = FALSE,
  gene_names = NULL
)
}
\arguments{
\item{df}{A dataframe with the first two MDS coordinates for a set of trees and distances.}

\item{consensus}{An optional argument specifying whether the first tree is a consensus tree, and
if so the name of the consensus tree.}

\item{group}{An optional variable to color points by.}

\item{title}{An optional paramater to change the plot title.}

\item{x_dim}{The name of the variable within the dataframe that specifies the value on the x axis.}

\item{y_dim}{The name of the variable within the dataframe that specifies the value on the y axis.}

\item{show_legend}{A boolean about whether or not to show a legend.}

\item{legend_lab}{A label for the legend (NULL if not specified).}

\item{scales_arg}{An optional argument about whether the scales should be fixed or free. Set to
free by default.}

\item{single_method}{True if the dataframe only includes data from a single distance metric.}

\item{gene_names}{An optional variable to label genes by.}
}
\value{
A ggplot2 object.
}
\description{
Plots MDS results
Plots MDS results of trees for each distance.
}
