% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_node_value.R
\name{plot_node_value}
\alias{plot_node_value}
\title{Plot node value
Plot a tree with node values on each internal node.}
\usage{
plot_node_value(
  main_tree,
  node_lab,
  n_trees,
  rooted = TRUE,
  lab_size = 2,
  supp_size = 2,
  xlim_max = 0.5
)
}
\arguments{
\item{main_tree}{A tree to check support for.}

\item{node_lab}{A list of node values for each internal node of main_tree.}

\item{n_trees}{The number of trees used to compute \code{node_lab}.}

\item{rooted}{A boolean, TRUE if the main tree is rooted}

\item{lab_size}{The size of the tip labels, default is 2.}

\item{supp_size}{The size of the support labels, default is 2.}

\item{xlim_max}{The length of the x axis.}
}
\value{
A ggtree object.
}
\description{
Plot node value
Plot a tree with node values on each internal node.
}
