% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_support.R
\name{plot_support}
\alias{plot_support}
\title{Plot support
Plot a tree with support values for each internal node. These can be gene tree 
support or bootstrap support values.}
\usage{
plot_support(
  main_tree,
  support,
  support_type = "gene",
  lab_size = 2,
  supp_size = 2,
  xlim_max = 0.5,
  hjust = 1,
  color_branch = FALSE,
  show_legend = TRUE,
  title = ""
)
}
\arguments{
\item{main_tree}{A rooted tree to plot edge information for.}

\item{support}{A list of values for each internal node of main_tree. Each value should be
a proportion between 0 and 1.}

\item{support_type}{Enter \code{"gene"} if \code{support} are gene tree support values,
\code{"boot"} if \code{support} are bootstrap support values, and \code{"other"} if 
\code{support} values are neither. Defaults to \code{"gene"}.}

\item{lab_size}{The size of the tip labels, default is 2. Set this to 0 to suppress
tip labels.}

\item{supp_size}{The size of the support labels, default is 2.}

\item{xlim_max}{The length of the x axis.}

\item{hjust}{Horizontal adjustment for support labels.}

\item{color_branch}{If TRUE, each branch is colored with the support of the internal node
it leads to. If FALSE, branch support is displayed with a number between 0 and 1.}

\item{show_legend}{A boolean about whether or not to show a legend for support.}

\item{title}{An optional parameter to change the plot title. Enter NULL to remove
the title.}
}
\value{
A ggtree object. Note this function will also provide a warning about rows
containing missing values if \code{color_branch = TRUE}. These rows correspond with
the branches that lead to tips, which all have 100% support.
}
\description{
Plot support
Plot a tree with support values for each internal node. These can be gene tree 
support or bootstrap support values.
}
\examples{
trees <- ape::rmtree(100, 5, rooted = TRUE)
support <- check_gene_support(trees[[1]], trees[2:100], rooted = TRUE)
plot_support(main_tree = trees[[1]], support = support, xlim_max = 2)

}
