% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne.R
\name{plot_tsne}
\alias{plot_tsne}
\title{Plots tSNE of log map results
Runs tSNE on the log map vectors and builds a low dimensional plot,
giving options to add colors and labels.}
\usage{
plot_tsne(
  vectors,
  phylogenomic = NULL,
  phylogenomic_name = NULL,
  group = NULL,
  other_tree = NULL,
  other_name = "other tree",
  title = "UMAP of Trees",
  show_legend = TRUE,
  legend_lab = NULL,
  tree_names = NULL,
  alpha = 1,
  trees_to_label = NULL,
  use_plotly = FALSE
)
}
\arguments{
\item{vectors}{A matrix with a row for the log map coordinate of each tree. The output
from compute_logmap.}

\item{phylogenomic}{An optional argument, if included gives the index of the \code{vectors} 
matrix that corresponds to the phylogenomic tree, which will then be noted in the plot.}

\item{phylogenomic_name}{An optional argument, gives the name of the \code{phylogenomic} tree.}

\item{group}{An optional vector to color points by. This should be same length as the 
number of rows in \code{vector}.}

\item{other_tree}{An optional additional tree to plot in a different color.}

\item{other_name}{An optional name for \code{other_tree} if it is included.}

\item{title}{An optional parameter to change the plot title.}

\item{show_legend}{A boolean about whether or not to show a legend for grouping variable.}

\item{legend_lab}{A label for the legend for the grouping variable. Enter "" to hide
the legend label.}

\item{tree_names}{An optional variable to label trees by. This should be same length as the 
number of rows in \code{vector}.}

\item{alpha}{Transparency of the points, defaults to 1 (fully opaque).}

\item{trees_to_label}{Optional set of names in \code{tree_names} to label on plot. These are
only labeled if \code{use_plotly} is set to \code{FALSE}.}

\item{use_plotly}{If true, the output will be a plotly object, with labels
when points are moused over, if false the output will be a ggplot object.}
}
\value{
A ggplot2 object, a dataframe of coordinates ordered by the component on the 
x-axis, and a dataframe of coordinates ordered by the component on the y-axis..
}
\description{
Plots tSNE of log map results
Runs tSNE on the log map vectors and builds a low dimensional plot,
giving options to add colors and labels.
}
\examples{
path <- paste0(system.file("txt", package = "groves"), "/")
lm_vectors <- compute_logmap(tree_paths = paste0(path, "tree", 1:4, ".txt"),
               tree_names = c("tree1", "tree2", "tree3", "tree4"))$vectors
names <- paste0("tree", 1:4)
med_branch <- c(1, 5, 2, 4)
plot_tsne(vectors = lm_vectors, phylogenomic = 1, phylogenomic_name = NULL, 
           group = med_branch, title = "UMAP plot", show_legend = TRUE, 
           legend_lab = "Branch median", tree_names = names,
           alpha = 0.9, use_plotly = FALSE, trees_to_label = "tree1")

}
