% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_long_branch.R
\name{recode_long_branch}
\alias{recode_long_branch}
\title{Recode long branches
Returns a phylo or multiphylo object with long branches recoded.}
\usage{
recode_long_branch(trees, branch_length)
}
\arguments{
\item{trees}{A phylo or multiphylo object.}

\item{branch_length}{A non-negative length to recode long branches to.}
}
\value{
A phylo or multiphylo object containing transformed trees with long branch
lengths recoded.
}
\description{
Recode long branches
Returns a phylo or multiphylo object with long branches recoded.
}
