% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_labs.R
\name{rename_labs}
\alias{rename_labs}
\title{Rename labels
Given a set of old and new labels, switch between the two.}
\usage{
rename_labs(match_df, tree, old_label = "oldlab", new_label = "newlab")
}
\arguments{
\item{match_df}{A dataframe with matched columns for the old and new labels.}

\item{tree}{A tree to rename tip labels for.}

\item{old_label}{The name of the column in match_df containing the old tip labels.}

\item{new_label}{The name of the column in match_df containing the new tip labels.}
}
\value{
The tree with renamed tip labels.
}
\description{
Rename labels
Given a set of old and new labels, switch between the two.
}
