% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Runs shiny app
Runs shiny app of groves with user supplied data or with an example Prevotella dataset.}
\usage{
run_shiny()
}
\value{
Nothing is returned.
}
\description{
Runs shiny app
Runs shiny app of groves with user supplied data or with an example Prevotella dataset.
}
