% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_MDS.R
\name{scale_MDS}
\alias{scale_MDS}
\title{Scales MDS results
Scales MDS results of trees for each distance so that all coordinates for each axis and distance
are between -1 and 1.}
\usage{
scale_MDS(df, n, group_var = "method", x_dim = "MDS1", y_dim = "MDS2")
}
\arguments{
\item{df}{A dataframe with the first two MDS coordinates for a set of trees and distances.}

\item{n}{The number of trees in the dataframe.}

\item{group_var}{The name of the variable within the dataframe that specifies the distance.}

\item{x_dim}{The name of the variable within the dataframe that specifies the value on the x axis.}

\item{y_dim}{The name of the variable within the dataframe that specifies the value on the y axis.}
}
\value{
A scaled version of the original dataframe.
}
\description{
Scales MDS results
Scales MDS results of trees for each distance so that all coordinates for each axis and distance
are between -1 and 1.
}
