% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_trees_txt.R
\name{write_trees_txt}
\alias{write_trees_txt}
\title{Writes trees as .txt files
Takes each tree in a phylo or multiPhylo object and writes it as a .txt file}
\usage{
write_trees_txt(tree, path, tree_names = NULL)
}
\arguments{
\item{tree}{The tree or set of trees to be written.}

\item{path}{The path to the folder to save the .txt files in.}

\item{tree_names}{A set tree names. If not given, trees will be numbered.}
}
\value{
A vector containing paths to the new tree .txt files
}
\description{
Writes trees as .txt files
Takes each tree in a phylo or multiPhylo object and writes it as a .txt file
}
\examples{
txt_path <- system.file("txt", package = "groves")
path <- paste0(txt_path, "/small_tree_set.txt")
tree_set <- ape::read.tree(path)
new_tree_set <- standardize_branches(tree_set)
write_trees_txt(tree = new_tree_set, path = paste0(txt_path, "/rescaled_small_trees"), 
                tree_names = paste0("tree", 1:3))

}
