% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments_to_cat.R
\name{alignments_to_cat}
\alias{alignments_to_cat}
\title{Makes new folder with alignments
Makes a new folder with alignments to concatenate into a single alignment.}
\usage{
alignments_to_cat(target_genes, path_from = "", tail = ".fa", path_to = "")
}
\arguments{
\item{target_genes}{A list of target gene names.}

\item{path_from}{Set to "".}

\item{tail}{Set to ".fa"}

\item{path_to}{Set to "".}
}
\value{
Nothing, creates a folder in the desired directory.
}
\description{
Makes new folder with alignments
Makes a new folder with alignments to concatenate into a single alignment.
}
