% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_missing_outgroup.R
\name{genes_missing_outgroup}
\alias{genes_missing_outgroup}
\title{Finds missing genes
Provides a subset of genes from a list that are missing in the outgroup.}
\usage{
genes_missing_outgroup(gene_names, presence, outgroup)
}
\arguments{
\item{gene_names}{List of genes of interest}

\item{presence}{Matrix of presence data for genes and genomes (TRUE if row gene is
present in column genome).}

\item{outgroup}{Name of the outgroup genome.}
}
\value{
A subset of gene_names that are missing in the outgroup genome
}
\description{
Finds missing genes
Provides a subset of genes from a list that are missing in the outgroup.
}
