% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_support.R
\name{plot_gene_support}
\alias{plot_gene_support}
\title{Plot gene tree support
Plot a tree with gene tree support values on each edge.}
\usage{
plot_gene_support(
  main_tree,
  n_trees,
  support,
  edge_second_node = (length(main_tree$tip.label) + main_tree$Nnode - length(support) +
    1):(length(main_tree$tip.label) + main_tree$Nnode),
  lab_size = 2,
  supp_size = 2,
  xlim_max = 0.5,
  hjust = 1
)
}
\arguments{
\item{main_tree}{A tree to plot edge information for.}

\item{n_trees}{The number of trees providing support for the main tree.}

\item{support}{A list of node values for each branch of main_tree.}

\item{edge_second_node}{Second node in the edge for each support value.}

\item{lab_size}{The size of the tip labels, default is 2.}

\item{supp_size}{The size of the support labels, default is 2.}

\item{xlim_max}{The length of the x axis.}

\item{hjust}{Horizontal adjustment for support labels.}
}
\value{
A ggtree object.
}
\description{
Plot gene tree support
Plot a tree with gene tree support values on each edge.
}
