% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logmap.R
\name{plot_logmap}
\alias{plot_logmap}
\title{Plot logmap
Plot logmap coordinates using the first two principal components.}
\usage{
plot_logmap(
  vectors,
  base_name,
  gene_names = NULL,
  col = NULL,
  col_name = NULL,
  show_legend = TRUE,
  cons_exists = TRUE
)
}
\arguments{
\item{vectors}{A matrix of logmap vectors for each tree.}

\item{base_name}{The name of the base tree in the log map.}

\item{gene_names}{An optional list of gene names. If no names are given, numbers will be used
to identify trees.}

\item{col}{An optional list of gene attributes to color the points of the scatterplot by.}

\item{col_name}{An optional name for the color attribute.}

\item{show_legend}{If true, the color legend is shown, if false it is hidden.}

\item{cons_exists}{True if plot includes a consensus tree, false if not.}
}
\value{
A ggplot object.
}
\description{
Plot logmap
Plot logmap coordinates using the first two principal components.
}
