% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logmap.R
\name{plot_logmap}
\alias{plot_logmap}
\title{Plots PCA of log map results
Plots PCA of log map results of trees for each distance.}
\usage{
plot_logmap(
  vectors,
  phylogenomic = NULL,
  group = NULL,
  title = "MDS of Gene Tree Distances",
  show_legend = TRUE,
  legend_lab = NULL,
  tree_names = NULL,
  alpha = 1,
  use_plotly = FALSE
)
}
\arguments{
\item{vectors}{A matrix with a row for the log map coordinate of each tree. The output
from compute_logmap.}

\item{phylogenomic}{An optional argument, if included gives the index of the \code{vectors} 
matrix that corresponds to the phylogenomic tree, which will then be noted in the plot.}

\item{group}{An optional vector to color points by. This should be same length as the 
number of rows in \code{vector}.}

\item{title}{An optional parameter to change the plot title.}

\item{show_legend}{A boolean about whether or not to show a legend for grouping variable.}

\item{legend_lab}{A label for the legend for the grouping variable. Enter "" to hide
the legend label.}

\item{tree_names}{An optional variable to label trees by. This should be same length as the 
number of rows in \code{vector}.}

\item{alpha}{Transparency of the points, defaults to 1 (fully opaque).}

\item{use_plotly}{If true, the output will be a plotly object, with labels
when points are moused over, if false the output will be a ggplot object.}
}
\value{
A ggplot2 object.
}
\description{
Plots PCA of log map results
Plots PCA of log map results of trees for each distance.
}
\examples{
path <- paste0(system.file("txt", package = "groves"), "/")
lm_vectors <- compute_logmap(tree_paths = paste0(path, "tree", 1:3, ".txt"),
               tree_names = c("tree1", "tree2", "tree3"))$vectors
names <- paste0("tree", 1:3)
med_branch <- c(1, 5, 2)
plot_logmap(vectors = lm_vectors, phylogenomic = 1, group = med_branch, title = "PCA plot",
         show_legend = TRUE, legend_lab = "Branch median", tree_names = names,
         alpha = 0.9, use_plotly = FALSE)

}
