% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MDS.R
\name{plot_MDS}
\alias{plot_MDS}
\title{Plots MDS results
Plots the results of MDS, giving options to add colors and labels.}
\usage{
plot_MDS(
  df,
  phylogenomic = NULL,
  phylogenomic_name = NULL,
  other_tree = NULL,
  other_name = "other tree",
  group = NULL,
  title = "MDS of Gene Tree Distances",
  show_legend = TRUE,
  legend_lab = NULL,
  tree_names = NULL,
  alpha = 1,
  use_plotly = FALSE,
  trees_to_label = NULL,
  x_axis = 1,
  y_axis = 2
)
}
\arguments{
\item{df}{A dataframe with the desired MDS coordinates for a set of trees and distances.
It should include variables named \code{MDS_x} and \code{MDS_y}.}

\item{phylogenomic}{An optional argument, if included gives the index of the dataframe that
corresponds to the phylogenomic tree, which will then be noted in the plot.}

\item{phylogenomic_name}{An optional argument, gives the name of the \code{phylogenomic} tree.}

\item{other_tree}{An optional additional tree to plot in a different color.}

\item{other_name}{An optional name for \code{other_tree} if it is included.}

\item{group}{An optional variable to color points by. This should either be a vector the same
length as the number of rows in \code{df} or a name in quotes of a variable in \code{df}.}

\item{title}{An optional parameter to change the plot title.}

\item{show_legend}{A boolean about whether or not to show a legend for grouping variable.}

\item{legend_lab}{A label for the legend for the grouping variable. Enter "" to hide
the legend label.}

\item{tree_names}{An optional variable to label trees by. This should either be a vector
the same length as the number of rows in \code{df} or a name in quotes of a variable in 
\code{df}.}

\item{alpha}{Transparency of the points, defaults to 1 (fully opaque).}

\item{use_plotly}{If true, the output will be a plotly object, with labels
when points are moused over, if false the output will be a ggplot object.}

\item{trees_to_label}{Optional set of names in \code{tree_names} to label on plot. These are
only labeled if \code{use_plotly} is set to \code{FALSE}.}

\item{x_axis}{Dimension number to use on the x-axis. Defaults to 1.}

\item{y_axis}{Dimension number to use on the y-axis. Defaults to 2.}
}
\value{
A ggplot2 object.
}
\description{
Plots MDS results
Plots the results of MDS, giving options to add colors and labels.
}
\examples{
trees_path <- system.file("txt", "small_tree_set.txt", package = "groves")
plot_df <- compute_MDS(trees_path = trees_path, tree_names = paste0("tree", 1:3))$df
plot_df$names <- paste0("tree", 1:3)
plot_df$type <- c("ribosomal", "ribosomal", "other")
plot_df$med_branch <- c(1, 5, 2)
plot_MDS(df = plot_df, phylogenomic = 1, group = "med_branch", title = "MDS plot",
         show_legend = TRUE, legend_lab = "Branch median", tree_names = "names",
         alpha = 0.9, use_plotly = FALSE)

}
