% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_outgroup.R
\name{recode_outgroup}
\alias{recode_outgroup}
\title{Recodes outgroup edge.
Recodes the branch going to an outgroup to the value provided.}
\usage{
recode_outgroup(tree, outgroup, value = NULL)
}
\arguments{
\item{tree}{The tree of interest.}

\item{outgroup}{The name of the outgroup in the tree of interest.}

\item{value}{The value to make the outgroup edge. If null, defaults to the second largest
length on the tree of interest.}
}
\value{
The tree of interest with the outgroup edge changed to the value provided.
}
\description{
Recodes outgroup edge.
Recodes the branch going to an outgroup to the value provided.
}
