% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_genomes.R
\name{remove_genomes}
\alias{remove_genomes}
\title{Removes genomes
Removes genome lines from a fasta file.}
\usage{
remove_genomes(target_genes, missing_genomes, path = "", tail = ".fa")
}
\arguments{
\item{target_genes}{A list of gene names for alignment files to remove empty lines from.}

\item{missing_genomes}{A list of genome names to remove from alignment files}

\item{path}{Path to the alignment files.}

\item{tail}{Ending of alignment files.}
}
\description{
Removes genomes
Removes genome lines from a fasta file.
}
