% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_branches.R
\name{standardize_branches}
\alias{standardize_branches}
\title{Standardizes branch lengths
Divides each branch length by the sum of branch lengths or maximum branch length
for that tree.}
\usage{
standardize_branches(tree, denom = "branch_sum")
}
\arguments{
\item{tree}{The tree or set of trees of interest.}

\item{denom}{The quantity to standardize by. Default is \code{branch_sum}, which is
the sum of branch lengths in that tree. The other option is \code{max_branch}, which
will divide each branch length by the maximum branch length in that tree.}
}
\value{
The standardized tree or set of trees.
}
\description{
Standardizes branch lengths
Divides each branch length by the sum of branch lengths or maximum branch length
for that tree.
}
\examples{
path <- system.file("txt", "small_tree_set.txt", package = "groves")
tree_set <- ape::read.tree(path)
new_tree_set <- standardize_branches(tree_set)

}
