% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Subsamples genes and genomes 
Subsamples genes and genomes to come up with a subset of genes and genomes such that
every genome contains every gene.}
\usage{
subsample(
  gene_names,
  path = "",
  tail = ".fa",
  genes_to_keep = NULL,
  genomes_to_keep = NULL,
  gene_weight = 1,
  genome_weight = 1
)
}
\arguments{
\item{gene_names}{A list of gene names that correspond with alignments to examine.}

\item{path}{The file path to the gene alignments, default is "".}

\item{tail}{The extension to the gene alignments, default is ".fa".}

\item{genes_to_keep}{An optional list of genes to keep in the subsample.}

\item{genomes_to_keep}{An optional list of genomes to keep in the subsample.}

\item{gene_weight}{Set to 1, the number of genes to add at once. If gene_weight is
larger than genome_weight, multiple genes will be added to the subsample before the next
genome.}

\item{genome_weight}{Set to 1, the number of genomes to add at once. If 
genome_weight is larger than gene_weight, multiple genomes will be added to the 
subsample before the next gene.}
}
\value{
presence: a matrix with rows for each gene and columns for each genome (tip). Entries 
equal to 1 mean the gene alignment includes that tip, and a list called
complete_set which includes a set of genes and genomes that were chosen as
the subsample.
}
\description{
Subsamples genes and genomes 
Subsamples genes and genomes to come up with a subset of genes and genomes such that
every genome contains every gene.
}
\examples{
gene_set <- paste0("gene_", 1:3)
faa_path <- paste0(system.file("faa/", package = "groves"), "/")
subsample(gene_names = gene_set, path = faa_path, tail = ".faa")

}
